% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{combineConditions}
\alias{combineConditions}
\title{Helper function to combine loaded data frames.}
\usage{
combineConditions(dfs, cond_names = names(dfs))
}
\arguments{
\item{dfs}{A list of data frames to combine}

\item{cond_names}{A character vector of strings that will be used to identify each
dataset.  If missing, the names of the list elements are used}
}
\description{
Takes a list of data frames, and names for those data frames, and combines
them using the data.table packages' rbindlist command and then adds a factor
variable (called `Condition`) with the name used to identify the data frame.
Useful for combining different datasets and analyzing them.
}
\examples{
# Using two arguments and a separate vector of names
ad = list(df1, df2)
nms = c("myFirst", "mySecond")
all = combineConditions(ad, nms)

# Using one argument and the names of variables
ad = list(myFirst = df1, mySecond = df2)
all = combineConditions(ad)
}

