% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fromPhred}
\alias{fromPhred}
\title{Function converts from Phred to error rate}
\usage{
fromPhred(qv, minError)
}
\arguments{
\item{qv}{The QVs associated with each value}

\item{minError}{The minimum allowed error rate.}
}
\value{
A vector of error rates
}
\description{
Function converts from Phred to error rate
}
\examples{
fromPhred(seq(10, 70, 10), 1e-5)
}

