% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getBAMNameFromDatasetFile}
\alias{getBAMNameFromDatasetFile}
\title{Get Aligned BAM Name}
\usage{
getBAMNameFromDatasetFile(dataset_name)
}
\arguments{
\item{filename}{The BAM file name}
}
\description{
This function parses a alignmentset.xml file and assuming it follows a "Standard"
format returns the path to the file name inside the file.
}
\examples{
getBAMNameFromDatasetFile("alignmentset.xml")
}

