% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getFastaFileNameFromDatasetFile}
\alias{getFastaFileNameFromDatasetFile}
\title{Get Reference Fasta File Path From Dataset}
\usage{
getFastaFileNameFromDatasetFile(dataset_name)
}
\arguments{
\item{filename}{The BAM file name}
}
\description{
This function parses a referenceset.xml file and assuming it follows a "Standard"
format returns the path to the file name inside the file.
}
\examples{
getFastaFileNameFromDatasetFile("referenceset.xml")
}

