% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getHoleX}
\alias{getHoleX}
\title{Convert a holeNumber into the X coordinate on the chip.}
\usage{
getHoleX(holeNumber)
}
\arguments{
\item{holeNumber}{the ZMW hole number}
}
\description{
Only applies to Sequel data where the hole number is
X << 16 & Y
}

