% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getHoleY}
\alias{getHoleY}
\title{Convert a hole number into the Y coordinate on the chip.}
\usage{
getHoleY(holeNumber)
}
\arguments{
\item{holeNumber}{the ZMW hole number}
}
\description{
Only applies to Sequel data where the hole number is
X << 16 & Y
}

