% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getReadReport}
\alias{getReadReport}
\title{Get Per Read Metrics}
\usage{
getReadReport(datasetname, indexedFastaName)
}
\arguments{
\item{indexedFastaName}{The fasta file used in the alignment.}

\item{dataset}{The dataset/BAM file name.}
}
\description{
This function loads a dataset and parses each read, passing it to
a class which collects metrics on each read.  At the end of he
}

