% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getReferencePath}
\alias{getReferencePath}
\title{Get the full path to a movie from a referenceset.xml file
and append the current path to it if it is a local reference.}
\usage{
getReferencePath(p)
}
\arguments{
\item{p}{A referenceset XML file.}
}
\description{
Get the full path to a movie from a referenceset.xml file
and append the current path to it if it is a local reference.
}

