% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{loadDataAtOffsets}
\alias{loadDataAtOffsets}
\title{Load BAM alignments as a list of data frames.}
\usage{
loadDataAtOffsets(offsets, bamName, indexedFastaName)
}
\arguments{
\item{offsets}{The virtual file offsets to retrieve BAM records from (can be obtained from the index file based on loadpbi).}

\item{bamName}{The BAM file name to grab}

\item{indexedFastaName}{The name of the indexed fasta file this should come from.}
}
\value{
Returns a list of alignments as data frames.  If the IPD and Pulse Width are available, they will be columns in the returned data as well.
}
\description{
Load BAM alignments as a list of data frames.
}

