% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{loadHeader}
\alias{loadHeader}
\title{Load BAM header}
\usage{
loadHeader(filename)
}
\arguments{
\item{filename}{The BAM file name}
}
\description{
This function loads the header of a BAM file into a List.  It will attempt to load the
list of sequences, program used and read groups as well as the version of the
BAM file, skipping over any information that is not present.
}
\examples{
loadHeader("~git/pbbam/tests/data/dataset/bam_mapping_1.bam")
}

