% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{loadHMMfromBAM}
\alias{loadHMMfromBAM}
\title{Load BAM alignments as a list of list for the HMM model.}
\usage{
loadHMMfromBAM(offsets, bamName, indexedFastaName, trimToLength = 140L)
}
\arguments{
\item{offsets}{The virtual file offsets to retrieve BAM records from (can be obtained from the index file based on loadPBI).}

\item{bamName}{The BAM file name to grab}

\item{indexedFastaName}{The name of the indexed fasta file this should come from.}

\item{trimToLength}{How much should we subsample the alignments?}
}
\value{
Returns a list of phase2datasets as data frames.
}
\description{
Load BAM alignments as a list of list for the HMM model.
}

