% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{loadpbi}
\alias{loadpbi}
\title{Load PBI BAM index file}
\usage{
loadpbi(filename)
}
\arguments{
\item{filename}{A character vector with the file name in it.}
}
\description{
This function loads a pbi index file into a dataframe.  Depending on the
number of attributes present, it will either load just the basic data, or optionally
the mapping and barcode data.
}
\examples{
loadpbi("~git/pbbam/tests/data/dataset/bam_mapping_1.bam.pbi")
}

