% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{loadReferenceWindow}
\alias{loadReferenceWindow}
\title{Load a section of the reference genome as a character string.}
\usage{
loadReferenceWindow(id, start, end, indexedFastaName)
}
\arguments{
\item{id}{The name of the sequence}

\item{start}{The start of the sequence}

\item{end}{The end of the sequence}

\item{indexedFastaName}{The name of the indexed fasta file this should come from.}
}
\value{
Returns a character vector of the sequence
}
\description{
Load a section of the reference genome as a character string.
}

