% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{loadRegionsTable}
\alias{loadRegionsTable}
\title{Load regions table}
\usage{
loadRegionsTable(subreadsBamName)
}
\arguments{
\item{subreadsBamName}{the .subreads.bam file name.  The corresponding .scraps.bam must be available in the same directory}
}
\value{
a data frame with columns HoleNumber, RegionType, RegionStart, RegionEnd, indicating the base extent of regions of each type.
}
\description{
Load a "regions table" for a subreads.bam (and optionally, a
corresponding `aligned_subreads.bam`).
}
\details{
Presently, this needs to traverse the entire BAM files---the PBI
doesn't contain any "region type" information.  If this becomes
crucial we might ' want to consider including it in the PBI.
}

