% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{loadSubreadsAtOffsets}
\alias{loadSubreadsAtOffsets}
\title{Load BAM subreads as a list of data frames.}
\usage{
loadSubreadsAtOffsets(offsets, bamName)
}
\arguments{
\item{offsets}{The virtual file offsets to retrieve BAM records from (can be obtained from the index file based on loadPBI).}

\item{bamName}{The BAM file name to grab}
}
\value{
Returns a list of subreads as a list of character strings.
}
\description{
Load BAM subreads as a list of data frames.
}

