% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getReadReport}
\alias{getReadReport}
\title{Get Per Read Metrics}
\usage{
getReadReport(datasetname, indexedFastaName)
}
\arguments{
\item{datasetname}{The dataset/BAM file name.}

\item{indexedFastaName}{The fasta file used in the alignment.}
}
\description{
This function loads a dataset and parses each read, passing it to a class
which collects metrics on each read.  It returns a list of data frames, one
for each metric analyzed.
}

