% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{loadReadsFromIndex}
\alias{loadReadsFromIndex}
\title{Given a data frame returned by loadPBI  load a list of the unaligned sequencing data.
(this is the unaligned equivalent of loadAlnsFromIndex)}
\usage{
loadReadsFromIndex(index, rows = NULL)
}
\arguments{
\item{index}{A data frame returned by loadPBI}

\item{rows}{A list of rows to select the data from (default = use all data)}
}
\value{
A list of alignments as data frames
}
\description{
Given a data frame returned by loadPBI  load a list of the unaligned sequencing data.
(this is the unaligned equivalent of loadAlnsFromIndex)
}
\seealso{
loadAlnsFromIndex
}

